Creating Boxplot for a dataset

The box and whiskers plot displays the distribution of a continuous variable. It visualises five summary statistics (the median, two hinges and two whiskers), and all "outlying" points individually. The `ggplot2` package is used here for creating the boxplot.

Boxplot_data(data)

Arguments

data

Proteomics expression dataset (original or normalized dataset)

Value

Interactive box and whiskers plot

Details

This can also be used for comparing the original dataset with the normalized dataset.

See also

Examples

Boxplot_data(yeast_data)
#> Using Majority protein IDs as id variables
#> Warning: Removed 269 rows containing non-finite outside the scale range
#> (`stat_boxplot()`).

Boxplot_data(rlr_knn_yeast_data)
#> Using Majority protein IDs as id variables