Creating Density plot for a dataset

Computes and draws kernel density estimate, which is a smoothed version of the histogram. This is a useful alternative to the histogram for continuous data that comes from an underlying smooth distribution. The `ggplot2` package is used here for creating the boxplot.

Densityplot_data(data)

Arguments

data

Proteomics expression dataset (original or normalized dataset) along with the protein information

Value

Interactive column-wise density plot

Details

This can also be used for comparing the original dataset with the normalized dataset.

See also

Examples

Densityplot_data(yeast_data)

Densityplot_data(rlr_knn_yeast_data)